feat: move BaseColbertLoss to own module
